server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
reportlab==4.2.2
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12